+++ core-ng-search-test/src/main/java/core/framework/search/impl/LocalNode.java
@@ -3,12 +3,11 @@ package core.framework.search.impl;
 import org.elasticsearch.analysis.common.CommonAnalysisPlugin;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
-import org.elasticsearch.index.mapper.MapperExtrasPlugin;
+import org.elasticsearch.index.mapper.extras.MapperExtrasPlugin;
 import org.elasticsearch.node.Node;
 import org.elasticsearch.painless.PainlessPlugin;
 import org.elasticsearch.plugins.Plugin;
-import org.elasticsearch.reindex.ReindexPlugin;
-import org.elasticsearch.transport.Netty4Plugin;
+import org.elasticsearch.transport.netty4.Netty4Plugin;
 
 import java.util.List;
 
@@ -17,11 +16,11 @@ import java.util.List;
  */
 class LocalNode extends Node {
     private static List<Class<? extends Plugin>> plugins() {
-        return List.of(ReindexPlugin.class,
-                Netty4Plugin.class,
-                MapperExtrasPlugin.class,  // for scaled_float type
-                PainlessPlugin.class,
-                CommonAnalysisPlugin.class);  // for stemmer analysis
+        return List.of(
+            Netty4Plugin.class,        // for http transport
+            MapperExtrasPlugin.class,  // for scaled_float type
+            PainlessPlugin.class,
+            CommonAnalysisPlugin.class);  // for stemmer analysis
     }
 
     LocalNode(Settings settings) {